General improvements to previous commit
Updated gitignore to exclude the database folder.

Updated the README

Updated: bot.load_modules, bot.update_challenge_board and bot.update_score_board to only affect one server.
If the same bot would run on multiple servers this should save resources when either one of the servers requests a reload of the bot.

Changed Event.challenges to be a dictionary where the key is the challenge name. This helps in performance when computing user scores

Changed how Event.solves works - Previously values in the list of solved challenges were of the form <Challenge Name><Server Id>. This is redundant and inhibits some functionality. Each Event object represents a different server, so the distinction in the solve challenge should not be required

Added the functionality to delete from the database entries of challenges no longer defined and entries of users no longer in the server.